docs(examples): add and fix remaining doc examples
