+++ demo/api.js
@@ -2,3 +2,6 @@ import './../apiExamples/alertValue.js';
 import './../apiExamples/error';
 import './../apiExamples/focus.js';
 import './../apiExamples/monthNames';
+import './../apiExamples/updateMaxDate';
+import './../apiExamples/updateMinDate';
+import './../apiExamples/validity';
